requirements.txt
Django~=4.0
python-dotenv~=0.19.2
PyJWT~=2.3.0
djangorestframework~=3.13.1